requirements.txt
pywinauto>=0.6.8
mss>=9.0.0
Pillow>=10.0.0
pyautogui>=0.9.54
PyQt6==6.7.1
PyQt6-Qt6==6.7.3
google-genai>=1.0.0
pynput>=1.7.6
